crypto-agent-main/crypto_agent/requirements.txt
chainlit
google-generativeai
aiohttp
python-dotenv
requests